feat: interrupt radio streams
Radio streams are infinite length tracks, so queuing requests after them is pointless.
If a radio stream is playing when the first song is added to the queue, then interrupt it
rather than queuing the selected track up next.

Tighten up the criteria for detecting the end of the playlist, so that "replace" can work
without triggering the default playlist. This is required for interrupting radio streams,
but it also improves the behaviour in other scenarios when MA is directly controlled to
override the default playlist.

- Instead of triggering whenever the queue length goes to 0, require it to go from 1 to 0
  This allows "replace" to be used when there is a playlist
- Additionally, require the queue length to remain at 0 for 2 seconds. This gives some
  time for the newly queued track to start even if the previous playlist was a single track.

This is not perfect. It works well when queuing local tracks, but sometimes tracks from a
network provider can take more than 2s to start and get replaced with the default playlist
when an internet radio station is the only thing in the playlist when queueing starts.